fix: address explainer PR review feedback
